+++ src/Symfony/Component/Serializer/Tests/DependencyInjection/AttributeMetadataPassTest.php
@@ -39,7 +39,7 @@ class AttributeMetadataPassTest extends TestCase
         (new AttributeMetadataPass())->process($container);
 
         $arguments = $container->getDefinition('serializer.mapping.attribute_loader')->getArguments();
-        $this->assertSame([false, []], $arguments);
+        $this->assertSame([false, [], []], $arguments, 'An empty discriminator map type registry should still be wired.');
     }
 
     public function testProcessWithTaggedServices()
@@ -70,7 +70,7 @@ class AttributeMetadataPassTest extends TestCase
             'App\Entity\Product' => ['App\Entity\Product'],
             'App\Entity\User' => ['App\Entity\User'],
         ];
-        $this->assertSame([false, $expectedClasses], $arguments);
+        $this->assertSame([false, $expectedClasses, []], $arguments);
     }
 
     public function testProcessWithForOptionAndMatchingMembers()
@@ -88,7 +88,7 @@ class AttributeMetadataPassTest extends TestCase
         (new AttributeMetadataPass())->process($container);
 
         $arguments = $container->getDefinition('serializer.mapping.attribute_loader')->getArguments();
-        $this->assertSame([false, [$targetClass => [$sourceClass]]], $arguments);
+        $this->assertSame([false, [$targetClass => [$sourceClass]], []], $arguments);
     }
 
     public function testProcessWithForOptionAndMissingMemberThrows()
@@ -104,6 +104,103 @@ class AttributeMetadataPassTest extends TestCase
             ->addTag('serializer.attribute_metadata', ['for' => $targetClass]);
 
         $this->expectException(MappingException::class);
+        (new AttributeMetadataPass())->process($container);
+    }
+
+    public function testProcessWithDiscriminatorMapType()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.source', _AttrMeta_DiscriminatorChild::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'child', 'discriminator_map_type' => true]);
+        $container->register('service.same_source', _AttrMeta_DiscriminatorChild::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'child', 'discriminator_map_type' => true]);
+
+        (new AttributeMetadataPass())->process($container);
+
+        $arguments = $container->getDefinition('serializer.mapping.attribute_loader')->getArguments();
+        $this->assertSame([false, [], [_AttrMeta_DiscriminatorParent::class => ['child' => _AttrMeta_DiscriminatorChild::class]]], $arguments);
+    }
+
+    public function testProcessCollectsSelfContributedDiscriminatorMapType()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.self', _AttrMeta_DiscriminatorParent::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'parent', 'discriminator_map_type' => true]);
+
+        (new AttributeMetadataPass())->process($container);
+
+        $arguments = $container->getDefinition('serializer.mapping.attribute_loader')->getArguments();
+        $this->assertSame([false, [], [_AttrMeta_DiscriminatorParent::class => ['parent' => _AttrMeta_DiscriminatorParent::class]]], $arguments);
+    }
+
+    public function testProcessRejectsDiscriminatorMapTypeOnUnknownClass()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.source', 'App\MissingChild')
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'child', 'discriminator_map_type' => true]);
+
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage(\sprintf('Class "App\MissingChild" cannot add a discriminator map type for "%s" because it cannot be found.', _AttrMeta_DiscriminatorParent::class));
+        (new AttributeMetadataPass())->process($container);
+    }
+
+    public function testProcessRejectsDiscriminatorMapTypeForUnknownTargetClass()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.source', _AttrMeta_DiscriminatorChild::class)
+            ->addTag('serializer.attribute_metadata', ['for' => 'App\MissingParent', 'type' => 'child', 'discriminator_map_type' => true]);
+
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage(\sprintf('Class "%s" cannot add a discriminator map type for "App\MissingParent" because the target class cannot be found.', _AttrMeta_DiscriminatorChild::class));
+        (new AttributeMetadataPass())->process($container);
+    }
+
+    public function testProcessRejectsUnknownTargetClass()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.source', _AttrMeta_Source::class)
+            ->addTag('serializer.attribute_metadata', ['for' => 'App\MissingTarget']);
+
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage(\sprintf('Class "%s" cannot extend serialization for "App\MissingTarget" because the target class cannot be found.', _AttrMeta_Source::class));
+        (new AttributeMetadataPass())->process($container);
+    }
+
+    public function testProcessRejectsDiscriminatorMapTypeThatIsNotASubtype()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.source', _AttrMeta_Source::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'source', 'discriminator_map_type' => true]);
+
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage('is not a subtype');
+        (new AttributeMetadataPass())->process($container);
+    }
+
+    public function testProcessRejectsDuplicateDiscriminatorMapTypes()
+    {
+        $container = new ContainerBuilder();
+        $container->register('serializer.mapping.attribute_loader', AttributeLoader::class)
+            ->setArguments([false, []]);
+        $container->register('service.first', _AttrMeta_DiscriminatorChild::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'child', 'discriminator_map_type' => true]);
+        $container->register('service.second', _AttrMeta_SecondDiscriminatorChild::class)
+            ->addTag('serializer.attribute_metadata', ['for' => _AttrMeta_DiscriminatorParent::class, 'type' => 'child', 'discriminator_map_type' => true]);
+
+        $this->expectException(MappingException::class);
+        $this->expectExceptionMessage(\sprintf('Discriminator map type "child" for "%s" is already mapped to "%s".', _AttrMeta_DiscriminatorParent::class, _AttrMeta_DiscriminatorChild::class));
         (new AttributeMetadataPass())->process($container);
     }
 }
@@ -129,4 +226,16 @@ class _AttrMeta_Target
 class _AttrMeta_BadSource
 {
     public string $extra;
+}
+
+class _AttrMeta_DiscriminatorParent
+{
+}
+
+class _AttrMeta_DiscriminatorChild extends _AttrMeta_DiscriminatorParent
+{
+}
+
+class _AttrMeta_SecondDiscriminatorChild extends _AttrMeta_DiscriminatorParent
+{
 }